# Fixed an issue with slow npm builds