--- a/package.json
+++ b/package.json
@@ -44,6 +44,7 @@
     "eslint-plugin-import": "^2.24.2",
     "eslint-plugin-react": "^7.26.1",
     "eslint-plugin-unused-imports": "^1.1.5",
+    "fork-ts-checker-webpack-plugin": "^7.0.0",
     "jest": "^27.3.1",
     "prettier": "2.3.2",
     "prop-types": "^15.7.2",