Enabled transpileonly for ts-loader to speed up npm builds.
(c.f. https://github.com/TypeStrong/ts-loader#transpileonly)

Removed inline-source-map because was able to debug it in the browser.
![image](https://user-images.githubusercontent.com/5561786/152753046-b93ec891-5593-416f-9c51-02b293f3168f.png)

- before
```
(.venv) dooga@hinagawa-dev:~/airone$ npm run watch

> airone@1.0.0 watch
> webpack build --watch --mode development

asset new-ui.js 10.9 MiB [emitted] (name: index)
1080 modules
webpack 5.36.2 compiled successfully in 19741 ms

asset new-ui.js 10.9 MiB [emitted] (name: index)
cached modules 3.49 MiB [cached] 1072 modules
runtime modules 2.25 KiB 6 modules
javascript modules 11.1 KiB
  ./frontend/src/AppRouter.tsx 6.17 KiB [built] [code generated]
  ./frontend/src/pages/Entity.tsx 4.9 KiB [built] [code generated]
webpack 5.36.2 compiled successfully in 4866 ms
```

- after
```
(.venv) dooga@hinagawa-dev:~/airone$ npm run watch

> airone@1.0.0 watch
> webpack build --watch --mode development

asset new-ui.js 5.28 MiB [emitted] (name: index)
1080 modules
webpack 5.36.2 compiled successfully in 7454 ms
Issues checking in progress...
No issues found.

asset new-ui.js 5.28 MiB [emitted] (name: index)
cached modules 3.5 MiB [cached] 1073 modules
runtime modules 2.25 KiB 6 modules
./frontend/src/pages/Entity.tsx 4.9 KiB [built] [code generated]
webpack 5.36.2 compiled successfully in 826 ms
Issues checking in progress...
No issues found.
```
Case with issue
```
asset new-ui.js 5.28 MiB [emitted] (name: index)
cached modules 3.48 MiB [cached] 1073 modules
runtime modules 2.25 KiB 6 modules
./frontend/src/components/entry/EntryForm.tsx 25.8 KiB [built] [code generated]
webpack 5.36.2 compiled successfully in 1088 ms
Issues checking in progress...
ERROR in ./frontend/src/components/entry/EntryForm.tsx 515:19-27
TS2322: Type 'number' is not assignable to type 'string'.
    513 |               <TableCell>
    514 |                 <EditAttributeValue
  > 515 |                   attrName={2}
        |                   ^^^^^^^^
    516 |                   attrInfo={attributes[attributeName]}
    517 |                   handleChangeAttribute={handleChangeAttribute}
    518 |                   handleNarrowDownEntries={handleNarrowDownEntries}
```